Cancel notification functionality
- Function to cancel notification from completed task/frog
- Moved setAlarm() from AlarmReceiver to Scheduler
