+++ app/src/main/java/com/metropolia/eatthefrog/notification/Scheduler.kt
@@ -1,10 +1,15 @@
 package com.metropolia.eatthefrog.notification
 
+import android.app.AlarmManager
+import android.app.PendingIntent
 import android.content.Context
+import android.content.Intent
 import android.util.Log
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.livedata.observeAsState
 import androidx.compose.ui.platform.LocalContext
+import com.metropolia.eatthefrog.activities.MainActivity
+import com.metropolia.eatthefrog.database.Task
 import com.metropolia.eatthefrog.viewmodels.NotificationsViewModel
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
@@ -66,7 +71,12 @@ fun Scheduler(viewModel: NotificationsViewModel) {
 /**
  * Function responsible for calling setAlarm() and sending the corresponding task for it
  */
-fun scheduleNotification(id: Long, option: String, viewModel: NotificationsViewModel, context: Context) {
+fun scheduleNotification(
+    id: Long,
+    option: String,
+    viewModel: NotificationsViewModel,
+    context: Context
+) {
     CoroutineScope(Dispatchers.IO).launch {
         val converter = DateTimeConverter()
         val task = viewModel.getCertainTask(id)
@@ -83,7 +93,46 @@ fun scheduleNotification(id: Long, option: String, viewModel: NotificationsViewM
                 setAlarm(task, modifiedTime.toString(), context)
             }
             setAlarm(task, task.time, context)
-        }
-        else setAlarm(task = task, context = context)
+        } else setAlarm(task = task, context = context)
     }
+}
+
+/**
+ * Function takes task as a parameter to use the tasks uid as a requestCode, which needs to be different for every alarm.
+ * The function will launch a notification when prompted and redirects user to MainActivity when the notification is clicked.
+ */
+fun setAlarm(task: Task, time: String = "09:00", context: Context?) {
+    val converter = DateTimeConverter()
+    val alarmManager = context?.getSystemService(Context.ALARM_SERVICE) as AlarmManager
+    val intent = Intent(context, AlarmReceiver::class.java)
+    intent.putExtra("task", task)
+
+    val pendingIntent =
+        PendingIntent.getBroadcast(context, task.uid.toInt(), intent, PendingIntent.FLAG_IMMUTABLE)
+    val mainActivityIntent = Intent(context, MainActivity::class.java)
+    val basicPendingIntent = PendingIntent.getActivity(
+        context,
+        task.uid.toInt(),
+        mainActivityIntent,
+        PendingIntent.FLAG_IMMUTABLE
+    )
+
+    val date: Date = converter.toTimestamp(time)
+    val now = Date()
+    Log.d("FUU date", date.toString())
+    Log.d("FUU date.time", date.time.toString())
+    Log.d("FUU now", now.toString())
+
+    if (date > now) {
+        val clockInfoTest = AlarmManager.AlarmClockInfo(date.time, basicPendingIntent)
+        alarmManager.setAlarmClock(clockInfoTest, pendingIntent)
+    }
+}
+
+fun cancelAlarm(task: Task, context: Context?) {
+    val alarmManager = context?.getSystemService(Context.ALARM_SERVICE) as AlarmManager
+    val intent = Intent(context, AlarmReceiver::class.java)
+    val pendingIntent =
+        PendingIntent.getBroadcast(context, task.uid.toInt(), intent, PendingIntent.FLAG_IMMUTABLE)
+    alarmManager.cancel(pendingIntent)
 }
